show snackbar after updating a product and move actions and effects into dedicated folders
