+++ apps/carbotracker/src/products-feature/+state/actions/routing.actions.ts
@@ -1,0 +1,8 @@
+import { createActionGroup, emptyProps } from '@ngrx/store';
+
+export const ProductsRouterActions = createActionGroup({
+  source: 'Products | Router',
+  events: {
+    'Navigated Away From Products Page': emptyProps(),
+  },
+});
