+++ apps/carbotracker/src/products-feature/+state/actions/snackbar.actions.ts
@@ -1,0 +1,8 @@
+import { createActionGroup, emptyProps } from '@ngrx/store';
+
+export const EditProductPageSnackBarActions = createActionGroup({
+  source: 'Products | Edit Product Page Snack Bar',
+  events: {
+    'Show Edit Product Snackbar Successful': emptyProps(),
+  },
+});
