+++ apps/carbotracker/src/products-feature/pages/EditProductPage/edit-product-page.component.ts
@@ -15,7 +15,7 @@ import {
   CtuiToolbarComponent,
 } from '@carbotracker/ui';
 import { Store } from '@ngrx/store';
-import { EditProductPageComponentActions as ComponentActions } from '../../+state/products.actions';
+import { EditProductPageComponentActions as ComponentActions } from '../../+state/actions/component.actions';
 import { productsFeature } from '../../+state/products.reducer';
 import { Product } from '../../product.model';
